+++ src/main/java/com/solvd/navigator/dao/mybatis/StreetImpl.java
@@ -1,7 +1,7 @@
 package com.solvd.navigator.dao.mybatis;
 
-import com.solvd.navigator.dao.IStreetLocationDao;
-import com.solvd.navigator.model.StreetLocation;
+import com.solvd.navigator.dao.IStreetDao;
+import com.solvd.navigator.model.Street;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.apache.logging.log4j.LogManager;
@@ -10,17 +10,17 @@ import org.apache.logging.log4j.Logger;
 import java.sql.SQLException;
 import java.util.List;
 
-public class IStreetLocationServiceImpl implements IStreetLocationDao {
-    private static final Logger LOGGER = LogManager.getLogger(IStreetServiceImpl.class);
-    private static final SqlSessionFactory SESSION_FACTORY = MyBatisUtil.getSqlSessionFactory;
-    StreetLocation streetLocation;
+public class StreetImpl implements IStreetDao {
+    private static final SqlSessionFactory SESSION_FACTORY = MyBatisUtil.getSqlSessionFactory();
+    private static final Logger LOGGER = LogManager.getLogger(StreetImpl.class);
+    Street street;
 
     @Override
-    public void insertEntity(StreetLocation entity) throws SQLException {
+    public void insertEntity(Street entity) throws SQLException {
         try (SqlSession sqlSession = SESSION_FACTORY.openSession()) {
-            IStreetLocationDao iStreetLocationDao = sqlSession.getMapper(IStreetLocationDao.class);
+            IStreetDao iStreetDao = sqlSession.getMapper(IStreetDao.class);
             try {
-                iStreetLocationDao.insertEntity(entity);
+                iStreetDao.insertEntity(entity);
                 sqlSession.commit();
             } catch (SQLException e) {
                 LOGGER.error("SQLException", e);
@@ -31,36 +31,36 @@ public class IStreetLocationServiceImpl implements IStreetLocationDao {
     }
 
     @Override
-    public StreetLocation selectEntityById(int id) throws SQLException {
+    public Street selectEntityById(int id) throws SQLException {
         try (SqlSession sqlSession = SESSION_FACTORY.openSession()) {
-            IStreetLocationDao iStreetLocationDao = sqlSession.getMapper(IStreetLocationDao.class);
-            streetLocation = iStreetLocationDao.selectEntityById(id);
+            IStreetDao iStreetDao = sqlSession.getMapper(IStreetDao.class);
+            street = iStreetDao.selectEntityById(id);
         } catch (SQLException e) {
             LOGGER.error("SQLException", e);
         }
-        return streetLocation;
+        return street;
     }
 
     @Override
-    public List<StreetLocation> selectAllEntity() throws SQLException {
-        List<StreetLocation> streetLocationList = null;
+    public List<Street> selectAllEntity() throws SQLException {
+        List<Street> streetList = null;
         try (SqlSession sqlSession = SESSION_FACTORY.openSession()) {
             try {
-                IStreetLocationDao iStreetLocationDao = sqlSession.getMapper(IStreetLocationDao.class);
-                streetLocationList = iStreetLocationDao.selectAllEntity();
+                IStreetDao iStreetDao = sqlSession.getMapper(IStreetDao.class);
+                streetList = iStreetDao.selectAllEntity();
             } catch (SQLException e) {
                 LOGGER.error("SQLException", e);
             }
         }
-        return streetLocationList;
+        return streetList;
     }
 
     @Override
-    public void updateEntity(StreetLocation entity) throws SQLException {
+    public void updateEntity(Street entity) throws SQLException {
         try (SqlSession sqlSession = SESSION_FACTORY.openSession()) {
-            IStreetLocationDao iStreetLocationDao = sqlSession.getMapper(IStreetLocationDao.class);
+            IStreetDao iStreetDao = sqlSession.getMapper(IStreetDao.class);
             try {
-                iStreetLocationDao.updateEntity(entity);
+                iStreetDao.updateEntity(entity);
                 sqlSession.commit();
             } catch (SQLException e) {
                 LOGGER.error("SQLException", e);
@@ -73,9 +73,9 @@ public class IStreetLocationServiceImpl implements IStreetLocationDao {
     @Override
     public void deleteEntity(int id) throws SQLException {
         try (SqlSession sqlSession = SESSION_FACTORY.openSession()) {
-            IStreetLocationDao iStreetLocationDao = sqlSession.getMapper(IStreetLocationDao.class);
+            IStreetDao iStreetDao = sqlSession.getMapper(IStreetDao.class);
             try {
-                iStreetLocationDao.deleteEntity(id);
+                iStreetDao.deleteEntity(id);
                 sqlSession.commit();
             } catch (SQLException e) {
                 LOGGER.error("SQLException", e);
